Refactor segmented control into a viewable
